Create Cart app
Create a new Cart app and move "Cart" and "CartProduct" models from Catalog to Cart
