+++ cart/admin.py
@@ -1,0 +1,12 @@
+from django.contrib import admin
+from . import forms
+from . import models
+
+
+@admin.register(models.Cart)
+class CartAdmin(admin.ModelAdmin):
+    form = forms.CartForm
+
+@admin.register(models.CartProduct)
+class CartProductAdmin(admin.ModelAdmin):
+    form = forms.CartProductForm
