+++ catalog/forms.py
@@ -10,13 +10,3 @@ class ProductForm(forms.ModelForm):
     class Meta:
         model = models.Product
         exclude = ['slug']
-
-class CartForm(forms.ModelForm):
-    class Meta:
-        model = models.Cart
-        exclude = ['total']
-
-class CartProductForm(forms.ModelForm):
-    class Meta:
-        model = models.CartProduct
-        exclude = ['subtotal']
