+++ catalog/models.py
@@ -1,8 +1,6 @@
 from django.db import models
 from django.utils.text import slugify
 
-from customer.models import CustomerProfile
-
 
 class SlugModel(models.Model):   
     name = models.CharField(max_length=200, unique=True, null=False, blank=False)
@@ -45,20 +43,3 @@ class Product(SlugModel):
 
     def __str__(self):
         return self.name
-    
-class Cart(models.Model):
-    products = models.ManyToManyField(Product, through="CartProduct")
-    customer = models.OneToOneField(CustomerProfile, on_delete=models.CASCADE, related_name='cart')
-    total = models.PositiveIntegerField(default=0)
-
-    def __str__(self):
-        return f"Panier de {self.customer}"
-
-class CartProduct(models.Model):
-    product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    cart = models.ForeignKey(Cart, on_delete=models.CASCADE)
-    number = models.PositiveIntegerField(default=1)
-    subtotal = models.PositiveIntegerField(default=0)
-
-    def __str__(self):
-        return f"{self.product} / {self.cart}"
